refactor: break prompt artist mix cycle
